+++ lib/migration_generator/migration_plus_generator.export.dart
@@ -4,3 +4,4 @@ export 'column_types.dart';
 export 'foreingkey_actions.dart';
 export 'foreingkey_ref.dart';
 export 'migration_plus_statement_generator.dart';
+export 'migration_plus_index_generator.dart';