+++ lib/migration_generator/statement_generator_base.dart
@@ -8,7 +8,7 @@ abstract class StatementGeneratorBase {
 
   StatementGeneratorBase(this.tableName, this.columns);
 
-  String getSqlStatement();
+  List<String> getSqlStatements();
 
   String getColumnSpec(ColumnDef col) {
     String spec = '';