+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.0
+
+** Breaking change
+
+* Refactoring the Alter Table class.
+
+  Previously it returns one string with several SQL statements. But the SQLite only runs one statement and stops.
+
+  Now it returns a list of string, so it runs all the statements, one at a time.
+
+* New CreateIndexGenerator class to easy the index creation
+
+
 ## 0.0.4
 
 * Renamed Migrator Plus Generator classes for easing code writting/reading